+++ CheckPoint/bios/choices.py
@@ -1,0 +1,5 @@
+SOURCE_CHOICES = [
+    ('Dumped from original hardware', 'Dumped from original hardware'),
+    ('Unknown source', 'Unknown source'),
+    ('Other', 'Other'),
+]