+++ CheckPoint/common/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from . import views
+from CheckPoint.common import views
 
 urlpatterns = [
     path('', views.home, name='home'),
@@ -20,4 +20,6 @@ urlpatterns = [
     path('screenshots/upload/', views.upload_screenshot, name='upload screenshot'),
     path('screenshots/my/', views.my_screenshots, name='my screenshots'),
     path('404/', views.custom_404, name='404 page'),
+    path('403/', views.custom_403, name='403 page'),
+    path('500/', views.custom_500, name='500 page'),
 ]