+++ CheckPoint/common/views.py
@@ -95,6 +95,14 @@ def custom_404(request, exception=None):
     return render(request, 'common/404.html', status=404)
 
 
+def custom_403(request, exception=None):
+    return render(request, 'common/403.html', status=403)
+
+
+def custom_500(request, exception=None):
+    return render(request, 'common/500.html', status=500)
+
+
 class CustomAdminLoginView(LoginView):
     template_name = 'accounts/login.html'